Let JSON classes emulate the types they wrap
By defining operators like __eq__, __gt__, etc, we enable easy interoperability of JSON instances and JSON-compatible values. More significantly, though, we can ensure: (a) equality of JSON instances in cases where the wrapped values are not equal but the JSON-equivalents are, for example an array [1,2,3] and a tuple (1,2,3); and (b) that we don't inadvertently consider falsey values equivalent to False and truthy values equivalent to True.