Enhance control over the application of subschemas
Keyword evaluation result is now optional, allowing some keywords (e.g. "then" in the case of "if" failing) to be skipped entirely. The keyword result now also has an assert_ flag, which may be toggled off to indicate that the result should not affect the outcome of the superschema (e.g. for the "if" keyword).