Change handling of keyword changes between JSON Schema versions
This effectively reverts 3df52eaf4e71f65a6eefc92056f284936ebdd658.

Spec'd keyword attributes like key name, allowed instance types, and keyword dependencies are once again fixed on keyword classes. There are no use cases in the transition from 2019-09 to 2020-12 in which a keyword class can simply be reused with a different key, so we don't need the overhead of dealing with key name mappings for referenced keywords, etc. The catalogue creation scripts are now beautifully simple again.

We deal with the deprecation of a keyword name/implementation by shunting it into vocabulary.legacy and appending the last version in which it works to its class name.