$dynamicRef - search scoped resources for $dynamicAnchor
Instead of looking for a "$dynamicAnchor" keyword in each schema in the
dynamic scope, look at anchors defined in each scoped schema *resource*.

The 2020-12 spec says `$dynamicRef` (when acting dynamically) replaces the
initially resolved URI with "the outermost schema resource in the
dynamic scope that defines an identically named fragment with
'$dynamicAnchor'."

The wording feels ambiguous, but the official test suite clarifies the
expectation:
Unlike the `$recursive*` keywords (which were looking for the "outermost
schema" that *contains* a specific keyword), `$dynamic*` are looking for
the outermost schema *resource* that *defines* a specific anchor.

tl;dr Official test suite cases were failing. The wording of the spec is
not great, but (as evidenced by this commit message) I can't find a
clearer way to word it.

Fixes #3