Add loadf/loads class methods for creating a JSON instance from JSON
And improve handling of floats in the constructor input: instead of converting directly to Decimal, we pass the float's string representation to the Decimal constructor. Consider that e.g. Decimal(0.1) => Decimal('0.1000000000000000055511151231257827021181583404541015625'), so previously we had JSON(0.1).value => Decimal('0.1000000000000000055511151231257827021181583404541015625'). Now, we effectively do Decimal(f'{0.1}'), which gives us JSON(0.1).value => Decimal('0.1').