Let scope validity depend transparently on its assertion behaviour
Previously we had (scope.valid => not scope.errors), but this was problematic for the "if" keyword, which must still pass even if it has errors. This was solved by testing scope._assert outside of the Scope class. In the case of "contains", in particular, things were getting increasingly hairy with temporarily suspending assertion behaviour, etc, etc.

Now we have (scope.valid => not scope._assert or not scope.errors), and we achieve the same behaviour for "then" and "else" simply by checking for errors on "if", which is now always valid. The "contains" implementation is simplified (actually, it's now the same as it was prior to the child scope indexing change). And, importantly from a class encapsulation perspective, the Scope._assert flag is now only referenced internally within Scope, as it should be.