Clarify the distinction between validation and assertion result
A new Scope.pass_ property returns the assertion result of a scope (as did Scope.valid, previously), while Scope.valid now returns the validation result.

For the currently supported JSON Schema vocabularies, validation and assertion result will only ever differ for the "if" keyword: its validation result may be false (triggering "else") while its assertion result is always true.

Schemas now check the pass_ property of their keywords' scopes in order to determine their own validity. Applicator keywords also (generally) check the pass_ property of the current, or child, scopes, in order to determine their own validity.

One consequence of this change is that output formatting code makes a bit more sense. Another consequence is that for the "basic" output format, if the overall validation result is false, errors for a failing "if" will now appear in the output.

This change does not affect an application's use of Scope.valid (as in the examples and docs) for determining the overall validation result, since a schema scope's validation and assertion results are always (in this implementation) identical.