Control scope validity independently of its error
This commit improves on 938ac95f87941b0dacd6c8692c743e42447ca7ee, further clarifying the nature of a Scope node.

Scope validity is no longer dependent on the presence of an error. This removes the need for a contrived error message in the case of schema (branch node) scopes, which allows these nodes to be easily left out of basic output, for example.