Fix "absoluteKeywordLocation" for by-reference keyword nodes
The absolute URI for a scope node is ordinarily based off the containing schema's URI. This works for child nodes of "$ref", "$dynamicRef" and "$recursiveRef" because we've already descended into the target schema by that stage. But a by-referenced keyword itself sits within some origin schema, and must set the target schema explicitly on its scope node. It can now do this using the `Scope.refschema` method.

Closes #15