Define the JSON-compatible type annotation as a type alias
It was defined as a type variable as part of the early idea of the JSON class as a generic. This may yet be explored in the future, but there are complications, e.g. with floats being converted to decimals, and in any case there are still a lot of issues (as shown by mypy) with type hinting in the json module. For now, a simple type alias is both more user-friendly and more correct.