Restrict hypothesis<6.0.4 due to a bug
The race condition fix for recursive() in hypothesis 6.0.4 has
broken the tests on Python 3.10 (and 3.11) due to the presence
of pytest-httpserver.  At least on an Apple M1 Pro chip.

Restrict it to 6.0.3 until the problem and solution are understood.
Note that all versions of pytest-httpserver seem to have the problem.

See the following issues on GitHub to track:

* csernazs/pytest-httpserver#217
* HypothesisWorks/hypothesis#3585