Fix bug in rel json ptr default value testing
This test needed to check against the string '0' rather than
the number 0.

Since the review feedback changed the code to
allow the empty string for the over parameter, this also reworks
the default value testing to handle any combination of parameters
that match their default values, including combinations with
the ref parameter as a JSONPointer instance.