Support xfail/skip of optional, format, next tests
This adds infrastructure to generate a JSON file flagging which
optional, format, and/or draft-next tests are expected to fail
("xfail") or have a bug that requires skipping them ("skip").
The file includes both the status to annotate the test, as
well as the reason.

The --testsuite-generate-status option will run all tests and
update the JSON file based on the current failures.  It will
leave in place existing entries for failures so as not to overwrite
and custom statuses or reasons.  It will remove entries for
tests that now pass.

Failing cases are set to "xfail" with an appropriate generic reason
(format, optional, or draft-next, in that order of precedence).
To skip a test, or provide a custom reason, edit the file manually.

This commit also includes documentation in in the CONTRIBUTING
file for test options I added a while back but apparently forgot
to document.