Fully unwrap annotation values
Both the annotation vocabulary and the unknown keywords annotation
class were using JSON.data for the annotation value.  This is fine
for scalar annotations, but not for object or array annotations like
the standard "examples", sometimes "default", and any non-scalar
unknown keyword annotations, as the array items or object values
are jschon.JSON-wrapped.

Annotations should either be fully jschon.JSON-wrapped or not wrapped
at all, so use JSON.value to preserve the scalar behavior and bring
objects and arrays into alignment.