Fine-grained JSON document structure control
This factors the instantiation of list items and object values
into their own methods to allow more flexible determination
of whether a sub-structure is a JSON Schema or not.  This will
allow using jschon with formats like OpenAPI that embed schemas
in complex pattern of locations, without the root object being
a schema.