Fix breakable test code
There's no need for the test pointer to be derived from the `value` parameter, which in this scenario is a (JSON-compatible) string, and for which Hypothesis found a breaking example: "~" is a valid JSON string, but "/~" is not a valid JSON pointer ("~" must be escaped). This raises a JSONPointerMalformedError instead of the expected JSONPointerReferenceError. Prior to 65314fafd0f228125f65038b7b6d404a29e3e094, any such invalid example went unnoticed, because the constructor just raised the JSONPointerError that was expected to occur during evaluation.